+++ rust/host/k16-vm/src/computer/devices.rs
@@ -2,6 +2,7 @@ mod bios;
 mod control;
 mod gpu;
 mod keyboard;
+mod mmu;
 mod serial;
 mod storage;
 mod timer;
@@ -12,6 +13,7 @@ pub(crate) use gpu::GpuDevice;
 pub(crate) use keyboard::validate_event as validate_keyboard_event;
 pub(crate) use keyboard::KeyboardDevice;
 pub use keyboard::KeyboardEvent;
+pub(crate) use mmu::{MmuControlCommand, MmuControlDevice};
 pub(crate) use serial::{DebugSerialDevice, SerialInputDevice};
 pub(crate) use storage::{
     K16VolumeFileStorageMedia, StoragePortControllerSnapshot, StoragePortDevice,